Traps better format OID

+++ pkg/snmp/traps/oid_resolver_test.go
@@ -24,8 +24,8 @@ var dummyTrapDB = trapDBFileContent{
 	},
 	Variables: variableSpec{
 		"1.3.6.1.2.1.2.2.1.1":      VariableMetadata{Name: "ifIndex"},
-		"1.3.6.1.2.1.2.2.1.7":      VariableMetadata{Name: "ifAdminStatus", Enumeration: map[int]string{1: "up", 2: "down", 3: "testing"}},
-		"1.3.6.1.2.1.2.2.1.8":      VariableMetadata{Name: "ifOperStatus", Enumeration: map[int]string{1: "up", 2: "down", 3: "testing", 4: "unknown", 5: "dormant", 6: "notPresent", 7: "lowerLayerDown"}},
+		"1.3.6.1.2.1.2.2.1.7":      VariableMetadata{Name: "ifAdminStatus", Enumeration: map[int64]string{1: "up", 2: "down", 3: "testing"}},
+		"1.3.6.1.2.1.2.2.1.8":      VariableMetadata{Name: "ifOperStatus", Enumeration: map[int64]string{1: "up", 2: "down", 3: "testing", 4: "unknown", 5: "dormant", 6: "notPresent", 7: "lowerLayerDown"}},
 		"1.3.6.1.4.1.8072.2.3.2.1": VariableMetadata{Name: "netSnmpExampleHeartbeatRate"},
 	},
 }
@@ -85,7 +85,7 @@ func TestDecoding(t *testing.T) {
 			"bar": VariableMetadata{
 				Name:        "yy",
 				Description: "dummy description",
-				Enumeration: map[int]string{2: "test"},
+				Enumeration: map[int64]string{2: "test"},
 			},
 		},
 	}
